+++ modules/runner/src/main/ts/reporter/Callbacks.ts
@@ -1,6 +1,15 @@
 import { ErrorData, Global } from '@ephox/bedrock-common';
-import Promise from '@ephox/wrap-promise-polyfill';
+
 import { HarnessResponse } from '../core/ServerTypes';
+
+export interface TestReport {
+  readonly file: string;
+  readonly name: string;
+  readonly passed: boolean;
+  readonly time: string;
+  readonly skipped: string | null;
+  readonly error: TestErrorData | null
+}
 
 export interface TestErrorData {
   readonly data: ErrorData;
@@ -11,15 +20,24 @@ export interface Callbacks {
   readonly loadHarness: () => Promise<HarnessResponse>
   readonly sendKeepAlive: (session: string) => Promise<void>;
   readonly sendInit: (session: string) => Promise<void>;
-  readonly sendTestStart: (session: string, totalTests: number, file: string, name: string) => Promise<void>;
-  readonly sendTestResult: (session: string, file: string, name: string, passed: boolean, time: string, error: TestErrorData | null, skipped: string | null) => Promise<void>;
+  readonly sendTestStart: (session: string, number: number, totalTests: number, file: string, name: string) => Promise<void>;
+  readonly sendTestResults: (session: string, results: TestReport[]) => Promise<void>;
   readonly sendDone: (session: string, error?: string) => Promise<void>;
 }
 
 declare const $: JQueryStatic;
 
-const sendJson = <T>(url: string, data: any): Promise<T> => {
+function generateErrorMessage(xhr: JQuery.jqXHR<any>, onError: (reason?: any) => void, url: string, requestDetails: string, statusText: 'timeout' | 'error' | 'abort' | 'parsererror', e: string) {
+  if (xhr.readyState === 0) {
+    onError(`Unable to open connection to ${url}, ${requestDetails}. Status text "${statusText}", error thrown "${e}"`);
+  } else {
+    onError(`Response status ${xhr.status} connecting to ${url}, ${requestDetails}. Status text "${statusText}", error thrown "${e}"`);
+  }
+}
+
+const sendJson = <T>(url: string, jsonData: any): Promise<T> => {
   return new Promise((onSuccess, onError) => {
+    const data = JSON.stringify(jsonData);
     $.ajax({
       method: 'post',
       url,
@@ -27,9 +45,9 @@ const sendJson = <T>(url: string, data: any): Promise<T> => {
       dataType: 'json',
       success: onSuccess,
       error: (xhr, statusText, e) => {
-        onError(e);
+        generateErrorMessage(xhr, onError, url, `sending ${data}`, statusText, e);
       },
-      data: JSON.stringify(data),
+      data,
     });
   });
 };
@@ -41,7 +59,7 @@ const getJson = <T>(url: string): Promise<T> => {
       dataType: 'json',
       success: onSuccess,
       error: (xhr, statusText, e) => {
-        onError(e);
+        generateErrorMessage(xhr, onError, url, 'as a get request', statusText, e);
       }
     });
   }));
@@ -64,8 +82,9 @@ export const Callbacks = (): Callbacks => {
     });
   };
 
-  const sendTestStart = (session: string, totalTests: number, file: string, name: string): Promise<void> => {
+  const sendTestStart = (session: string, number: number, totalTests: number, file: string, name: string): Promise<void> => {
     return sendJson('/tests/start', {
+      number,
       totalTests,
       session,
       file,
@@ -73,15 +92,10 @@ export const Callbacks = (): Callbacks => {
     });
   };
 
-  const sendTestResult = (session: string, file: string, name: string, passed: boolean, time: string, error: TestErrorData | null, skipped: string | null): Promise<void> => {
-    return sendJson('/tests/result', {
+  const sendTestResults = (session: string, results: TestReport[]): Promise<void> => {
+    return sendJson('/tests/results', {
       session,
-      file,
-      name,
-      passed,
-      skipped,
-      time,
-      error,
+      results,
     });
   };
 
@@ -101,7 +115,7 @@ export const Callbacks = (): Callbacks => {
     sendInit,
     sendKeepAlive,
     sendTestStart,
-    sendTestResult,
+    sendTestResults,
     sendDone
   };
 };