+++ CHANGELOG.md
@@ -4,6 +4,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 15.0.1 - 2025-01-27
+
+## Fixed
+- AWS remote testing idle timeout was not set
+
+## 15.0.0 - 2025-01-27
+
+## Added
+- New server-side APIs to accept a batch of results instead of a single result #TINY-11177
+
+## Changed
+- Reverted TINY-10708 which was a server-side fix
+- Client no longer waits for log requests to complete between tests, which should speed up remote testing #TINY-11177
+- Console HUD no longer updates for individual tests #TINY-11177
+- Client now posts test status only in batches every 30 seconds, this is the only time the console HUD will update #TINY-11177
+- Server now controls the remote webdriver idle timeout, and makes sure a keep-alive command is sent regularly
+
+## Removed
+- Single result server-side API #TINY-11177
+- Server-side monitoring of single test timeouts. This is still monitored client side. #TINY-11177
+- The Promise polyfill is no longer allowed on modern NodeJS frameworks so it has been removed. #TINY-11177
+
+## 14.1.5 - 2024-10-18
+
+### Added
+
+- Added static mock service worker js file mapping. #TINY-11415
+
 ## 14.1.4 - 2024-03-27
 
 ### Fixed